tests/mediacapture-depth/BAK/MediaStreamTrack-MediaElement-disabled-video-is-black-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>A disabled video track is rendered as blackness</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#life-cycle">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<h2>Description</h2>
<p>
  This test checks that a disabled video track in a
MediaStream is rendered as blackness.
</p>

<h2>Preconditions</h2>
<ol>
  <li>
    Use a test device with depth camera(embedded or external).
  </li>
  <li>
    When prompted, accept to share your depth stream.
  </li>
</ol>

<div id="log"></div>
<script src="../common/vendor-prefix.js" data-prefixed-prototypes='[{"ancestors":["HTMLMediaElement"],"name":"srcObject"}]'></script>
<video id="vid"></video>
<script>

let vid = document.getElementById("vid");
let cv = document.createElement("canvas");
async_test(t => {
  return navigator.mediaDevices.getUserMedia({video: {videoKind: {exact: "depth"}}})
  .then(stream => {
    if (stream.getVideoTracks()[0].enabled) {
      stream.getVideoTracks()[0].enabled = false;
    }

    let testOnceLoadeddata = function() {
      vid.removeEventListener("loadeddata", testOnceLoadeddata, false);
      cv.width = vid.offsetWidth;
      cv.height = vid.offsetHeight;
      let ctx = cv.getContext("2d");
      ctx.drawImage(vid,0,0);
      let imageData = ctx.getImageData(0, 0, cv.width, cv.height);
      for (let i = 0; i < imageData.data.length; i+=4) {
        assert_equals(imageData.data[i], 0, "No red component in pixel #" + i);
        assert_equals(imageData.data[i + 1], 0, "No green component in pixel #" + i);
        assert_equals(imageData.data[i + 2], 0, "No blue component in pixel #" + i);
        assert_equals(imageData.data[i + 3], 255, "No transparency in pixel #" + i);
      }
      t.done();
    }

    vid.srcObject = stream;
    vid.play();
    vid.addEventListener("loadeddata", t.step_func(testOnceLoadeddata), false);
  });
}, "Tests that a disabled video track in a MediaStream is rendered as blackness");

</script>
